+++ src/domain-model/service-registration.js
@@ -1,4 +1,5 @@
 import { ServiceCollection } from '../foundation/di/service-collection.js';
+import { AggregateRegistry } from './aggregates/aggregate-registry.js';
 import { BoundedContextRegistry } from './bounded-contexts/bounded-context-registry.js';
 import { CoreEntityRegistry } from './entities/core-entity-registry.js';
 import { DomainOverviewDescriptor } from './overview/domain-overview-descriptor.js';
@@ -9,6 +10,7 @@ export function addDomainModel(services) {
     throw new Error('addDomainModel expects an instance of ServiceCollection.');
   }
 
+  services.registerSingleton('AggregateRegistry', () => new AggregateRegistry());
   services.registerSingleton('BoundedContextRegistry', () => new BoundedContextRegistry());
   services.registerSingleton('CoreEntityRegistry', () => new CoreEntityRegistry());
   services.registerSingleton('DomainOverviewDescriptor', () => new DomainOverviewDescriptor());